Chore: 2주차 폴더 생성
